Execute multiple statements, including non-SQL statements.
The previous code assumed the BODY only ever contained SQL statements
terminated by a semi-colon, did rudimentary parsing and splitting,
folded each statement to a single line, and executed them.

This worked -- roughly -- but meant that ob-sql-mode could not issue
non-SQL statements. For example, in the SQLite shell the output
format can be changed by issuing ".mode ..." commands, which do not
end with a semi-colon.

Change this, and send the entire buffer to the shell process. Then
post-process the results and remove any text that matches the shell's
prompt or continuation prompt.